pyproject.toml
[project]
name = "lilith"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl_cffi>=0.7.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "docker>=7.1.0",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "python-telegram-bot>=22.6",
    "markdown-it-py>=4.0.0",
    "bleach>=6.0.0",
    "pydantic>=2.0.0",
    "mcp>=1.0.0",
    "langsmith>=0.2.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["migrations", ".venv", "venv", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4", "TCH", "PTH", "N", "W"]
ignore = ["E501", "E701", "E402", "F821", "F841", "W291", "B007", "B904", "B905", "SIM102", "SIM103", "SIM105", "SIM115", "PTH123", "TC003"]

[tool.ruff.format]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
